requirements.txt
Django~=2.2.1
django-crontab==0.7.1
djangorestframework=3.9.4
google-api-python-client=2.15.0